[#601] urls: move urls.py to individual apps
