+++ cove360/urls.py
@@ -1,0 +1,5 @@
+from django.conf.urls import url
+
+from cove import urlpatterns
+
+urlpatterns += [url(r'^data/(.+)$', 'cove360.views.explore', name='explore')]
