+++ coveIATI/urls.py
@@ -1,0 +1,6 @@
+from django.conf.urls import url
+
+from cove import urlpatterns
+
+
+urlpatterns += [url(r'^data/(.+)$', 'coveIATI.views.explore', name='explore')]
