+++ coveOCDS/urls.py
@@ -1,0 +1,6 @@
+from django.conf.urls import url
+
+from cove import urlpatterns
+
+
+urlpatterns += [url(r'^data/(.+)$', 'coveOCDS.views.explore', name='explore')]
